mission_basic: polygon objects in distance sensor
Made it possible to define objects based on points that make use
a polygon, optionally with differing altitudes as well. These polygons
can be placed in cardinal directions but also differently if wanted.

The distance sensor can now calculate the distances to any of the edges
defined by these polygons using ray casting. First, the vehicle yaw is
checked to be within the minimal and maximal angle of the polygon from
the vehicle location (with some modulo and quadrant tricks). Then we
calculate the parameters (`m` and `b` in `y = mx + b`) that make up the
lines of an edge and the ray of the vehicle's yaw angle. This allows us
to find the point of intersection (using some deductions to simplify the
calculation a bit). Finding the altitude and distance of that point is
then quite simple.