Sensor: Detect whether we flew into an object
Sometimes the sensor is not questioned often enough causing us to fly
into the object. Usually, the edges are then detected when flying out of
the object again, but we should objectively detect whether the vehicle
is inside the object in any case. This is what the ray insersection code
is for, with some additions to keep the altitude in mind (simplified).