Remove dronekit LICENSE and update README.md
The dronekit example scripts are licensed under the Apache 2.0 license,
but this is not a problem since it is compatible with GPLv3. That means
we simply distribute it under the latter license. It is recommended by
both the Apache Foundation and the FSF to use one license (the more
restrictive one, in this case GPLv3) instead of dual-licensing in this
case. See http://www.apache.org/licenses/GPL-compatibility.html for
details.

Briefly mention this in the README for future reference.