Imported documentation
These documents were previously placed elsewhere, but it is useful to
have it all in one place. These are installation instructions for
ardupilot/dronekit and the Raspberry Pis, as well as investigations and
references for parts of the project. We could later merge information
from these into the main README.

Included a readme about this and made git ignore compiled files. Added
a makefile to easily compile the PDFs.