Settings: Store component data associatively
This simplifies the lookup for the component name in the settings data.
Since we only ever need one settings object for each component, we can
store the component identifiers associatively.

Add a descriptive (printable) name of the component in the object as
well, which might be useful for help messages, etc. Thus the structure
of the object itself is mostly the same and the abstracted logic of
Settings means that nothing else needs to be changed.