Added ArduPilot files .gitignore
We have currently described two ways of running the ArduPilot simulator:
from the vehicle directory within ardupilot (docs/Installation_Leon.txt)
or from the directory with the Python script (docs/Installation_Tim.tex)
-- both work, although the first option requires giving relative paths
to the actual script, which is somewhat tedious but can be smoothened by
using a mission script to run the commands.

We are soon required to run the ArduPilot binary from within the same
directory as the Python mission script (the second option), because we
want to have a simple mission command script tracked in Git as well.
This means the paths in the commands are relative to this repository,
which means we should use the first option.

ArduPilot compiles binaries and makes log files. Ignore those.