Move mission script to main directory
Temporarily place the mission script in the main directory. The classes
in the Python script will be moved into separate modules later on. This
is required to use the settings component later on -- using the settings
in the other components requires some packaging work as well.

Also added the mission command script, which is run in ArduPilot when it
is ready. The documentation for this is scattered around a bit, but will
be merged to the README later on.