Settings: Simplify import of class
The settings module currently only contains a Settings.py file with
a Settings class. Most likely we will only need that class anyway, and
other classes would either be in additional files or should be "private"
classes. Thus we can import the class directly by importing it in the
__init__ script, which nicely packages it up.

This makes the line `from settings.Settings import Settings` a bit
cleaner, namely `from settings import Settings`. This will also be very
helpful for cross-component imports.

This was tested with the XBee sensor simulator to work.