mission_basic: Make use of settings
This replaces some magical constants with settings that can be changed
in the JSON file, as well as adds some documentation (in the code) for
the setting variables. There are still some hardcoded parameters left
that would be difficult or currently useless to change into settings,
but those could be done later on.