Settings: Keep loaded JSON object in class
Before now, each instantiation of Settings will open the JSON file,
parse it and take one of the components from it. In the case that
a script has multiple users of Settings, such as mission_basic.py with
the 'distance_sensor_simulator' and the 'mission' itself, this means
that the file is being opened multiple times. Later on, we might in fact
use the XBee component from the trajectory mission as well.

Therefore, keep the loaded object in a static variable and add a class
method to load it. Since the Settings object can be given a different
JSON file each time, use a dictionary so that they don't collide with
each other.