mission_basic: Separate geometry utility functions
Move the utility functions into a utils package in a Geometry module
file. These functions can be used by classes from different components,
so it is useful to keep them in a central place. It is also easy to
split off since it has very few dependencies, at least none within the
toolchain.

Added a README file about the purpose and a notice that code may be
based on e.g. DroneKit examples.