mission_basic: Tweak mission settings
Instead of retrieving the Settings in the Mission class itself and then
relying on Mission.settings in the main function, load the settings in
main, pass it to the Mission class and use it later on as well. This is
more in line with the other scripts right now, and even though the main
code might be moved into the Mission class as well it seems nicer to do
it this way.

On the other hand, the two distance sensors now have separate components
in the settings file. For those it might make more sense to let the
class handle the import by itself, or we should merge those components
in an effort to streamline the interfaces of the classes for inheritance
(leaving the physical implementation with simulation settings, however).