mission_basic: Move classes to packaged components
Move the Mission class back into the trajectory directory, and move the
Sensor class into a distance component, with the class renamed to
Sensor_Simulator as a step to make the physical sensor base off of
a similar interface.

To simplify/allow the imports in mission_basic, use __init__ scripts to
import the files so that we can quickly import the classes themselves.

The main problem with putting code into separate packages in Python is
that when such code needs to import something from another subpackage,
it needs to use a relative import. These are only allowed however when
the parent directory is also a package, and this package must be
correctly initialized on runtime. When we run a script like
mission_basic.py, however, there is no package and we just run in some
global scope like __main__. Thus we need to initialize the package.

There are multiple boilerplate approaches to do so, but to keep this
centralized the initialization actually happens in __init__, which seems
logical, but the problem is that __init__ is only loaded if the code
already thinks it's in a package... Thus we load the package in the
following way:
from __init__ import __package__
This makes the script being run apply the package name to itself, and
then the imports of subpackages work correctly. This line should thus be
added before any subpackage import, at least if that subpackage has
relative imports.