Added initial memory map code
This currently creates a map of the nearby surroundings using the
distance sensor to detect where a point could be. It is still very
simplistic and depends hugely on the mission and simulated objects.
Right now we visualize it using a plot that also shows the "real"
locations of the object as well as the edges that are relevant to the
measured distance.