Initial work on importing VRML scenes
This makes it possible to parse a VRML scene using OpenGLContext's vrml
package. The scene graph is then traversed to find nodes with geometry
coordinates which can then be converted to Locations for the Environment
which uses the VRMLLoader to make simulated objects that are not
hardcoded.

This does not yet use the translation information from the file, and
most VRML files contain so many faces (shapes) that although parsing is
fast, the distance sensor calculations are not. The calculations
themselves understand the format of the faces but do not yet handle the
planes themselves correctly, where they only search for edge distances
rather the directional distance.