Initial work on different geometry systems
For simplified simulations, we want to step away from the use of
geographic coordinates and use the raw meters instead. This means we now
have a geometry object within the environment, which knows what kind of
coordinate system we use. This also changes/adds some accessors in
Environment to be more useful.

The new non-spherical system does not work very well yet since some
locations still use coordinates, which causes the vehicle to not change
position in the memory map (which is not drawn correctly). The Mission
still needs to use spherical anyway since it interfaces with dronekit
directly. The non-spherical system can only be used if the trajectory
simulation does not happen within dronekit/ardupilot.