Added MockVehicle as a replacement for dronekit
This allows running mission_basic standalone instead of in the ardupilot
simulator. The Mission class has been updated to change data in the mock
class if it is used, and now uses the environment's geometry. It is
recommended to set the geometry_class to 'Geometry' for the mock script
and to 'Geometry_Spherical' when running the script in ardupilot.

The MockVehicle can be given settings and has accessors for its fields
as expected, so very little needs to be changed elsewhere aside from
Mission. The MockVehicle does not yet update its location and so on, but
setting the yaw does work.