MockVehicle: Make it possible to fly
The MockVehicle can now at least fly up, and flying around should also
work with AUTO mode commands or GUIDED velocity/goto commands. Whether
all the directions actually work is not yet tested, but changing the yaw
is supported (if only that it is still instantaneous, unlike the speed).

Also simplified Geometry.get_angle to use atan2 instead of atan with
quadrant fixes. Pass the Geometry to the MockVehicle so it can use it
for calculations.