MockVehicle: More work on AUTO mode commands
Go to next command if altitude/location reached and correctly fly to the
exact destination altitude. Whether this also needs to be done for
locations is not yet relevant since the mission should accommodate for
reaching imprecise waypoint, except that distance_to_current_waypoint
does not yet work because it uses the wrong command for some reason.