Attempt to improve memory map plot speed
After a while, the performance of the displayed memory map dropped which
stalled the entire loop. This seems to because the old plot data is
still kept and redrawn with another image. Clear the plot after drawing
it so that this data is discarded. This also means we no longer have to
remove the annotation line of the relevant edge since this is done
automatically.