Distance_Sensor_Simulator: Track edge object
Previously the edge of some other object could be chosen as the relevant
object. Pass the edge along for each object distance measurement so that
we can eventually keep the one with the closest distance.

There is some initial support for plotting non-edge relevant objects,
namely the center of a circular object, but this is not very
comprehensive yet.