Geometry: Fix ray intersection angles
The algorithm for ray intersection of segments incorrectly attempted to
use the angle calculation when the longitudes are less than epsilon
apart, which could potentially give a division by zero error, while
longitudes that were far apart had their angle set to the float maximum.

This was the wrong way around, also considering the code it was based
on. This is now corrected, which should make the memory map "cheat"
display better.