mission_basic: Use memory map cheat when no scene
The logic for when to use the "cheat" of showing where the simulated
objects are was the wrong way around: instead of not showing it when we
use a VRML scene file (which potentially has lots of objects), we did
show it in that case and were not showing it in the case of the default
hardcoded objects. This logic is now inverted correctly.