mission_basic: Direct collision data cheat
Use the polygon data itself to display the collision data on the memory
map. This means that it is no longer necessary to check each point
whether it is inside the polygon or not, and makes the display much
faster. There are no more lines where the ray casting algorithm failed
to move the point off the polygon edges, and gives a clear and correct
view of the locations of the object and its edges.