Distance_Sensor_Simulator: Fix points off edges
When the vehicle's angle is in such a way that it could potentially see
the object based off from the angles/quadrants alone, we proceed to
detect the intersection of the ray and any of the edges of the object.
However, some points of intersection may not lie on the edges
themselves, but on lines extending from them. To prevent those points to
be considered for the minimal distance, skip them when the distance to
both edge points is greater than the distance between the two edge
points themselves. This works to counter incorrect distance
measurements.

Also speed up one additional call to min.