Memory_Map: Fix polygon and detected edge drawing
For any positioning other than displaying an image, matplotlib assumes
the value is given in (x,y) order instead of (y,x). This causes the
wrong relevant edge to be drawn as well as the polygons themselves to be
misplaced.

Also draw the "cheat" edges/polygons using alpha transparency so that
the actual memory map with detected points is still mostly visible.