Settings: Support parent component inheritance
The XBee components contained a duplicate setting for number_of_sensors.
It seems tedious and error-prone to change all instances of them when
it's mostly only related to the main component right now, the XBee
simulator. Make it possible to transparently get a setting from
a registered parent component to avoid this issue.

Later on the components might even become more hierarchical with base
components for simulated and physical instances of sensors, etc. This
will be helpful in that case, as well.

Added tests to demonstrate that it works.