Settings: Add Arguments registration class
This class registers Settings objects so that we instantiate only one
object per component, and also registers argument specifications in the
argument parser. This allows the Settings objects to receive overridden
values via command line input, which is helpful to perform temporary
tests with values without having to change the JSON file.

Additionally, we can now pass a settings file as the very first
positional argument (without an --option name in front) to use
a different JSON file for all the Settings objects.

The check for help messages has to be delayed until all the Settings
objects are registered and all the known arguments are parsed, which is
a bit messy but options are indeed shown in groups.