XBee: Use Arguments to read command line settings
Use the Arguments class to retrieve inputs from the command line and to
store Settings objects that would otherwise be created multiple times.
This moves the default JSON file to one location in the script and
required fairly little interference in the classes.

The only problem is that the XBee Viewer shouldn't start a plot
immediately upon initialization, because then we can't get argument help
without opening a nonfunctional plot viewer. Therefore delay starting
the plotting when we actually draw points on it. We potentially show
help after all objects are created, which may be more than absolutely
necessary but should still work fine.