Make the XBee configurator work for all sensors and use settings
Reads the number of sensors `n` from the settings file and lets the user
configure `n + 1` sensors as the ground station also needs to be
configured and is left apart from the total sensor count.

Additionally we make use of the settings component to store and fetch
the settings for this component.