Add XBee sensor configuration documentation
These notes capture the steps required to setup the sensors before we
can do the rest of the setup and controlling with Python.

Additionally remove a comment from the Raspberry Pi documentation as no
extra package was necessary: `bsdtar` comes along with Arch Linux ARM
according to the installation guide.