mission_basic: Fix problems with missing polygons
If we're not in a simulation, a scenefile is used or no hardcoded
objects are around, there will be no polygons shown on the memory map,
so the "patches" collection will be empty. Apparently matplotlib cannot
handle that well and gives a confusing error. Skip adding it at all in
this case.