Distance_Sensor_Simulator: Support 3D polygons
This allows the distance sensor to detect the objects from VRML files
how they are supposed to, i.e. detecting any point in the polygon rather
than just the edges. While we still assume the distance sensor always
detects on the same level, we need to be able detect more complex
polygons now.

We use several 3D algorithms to convert the polygon to a plane, find the
intersection point on the plane, project the plane into 2D and use the
regular point-in-polygon check to find if the intersection is actually
within the polygon. This seems to work acceptably, but has not been
fully tested.

Also cleaned up some other code.