Switch to unicasting instead of broadcasting
It turns out that broadcasting is really bad for performance in ZigBee
networks due to huge overhead when transferring the message. Unicasting
(directly addressing the destination sensor) is much more efficient and
by using unicasting we completely resolve the huge delays we experienced
before.

This patch additionally simplifies the code in the runner a bit and
moves TODO statements to the top of the file for overview.