Unify XBee sensor APIs
Make the `_receive()` method take a received packet as parameter. This
was already the case for the physical sensor class, but now the
simulated sensor class has this too, as well as the base class.

This patch also implements the `activate()` method for the physical
sensor to avoid accessing a private method in the runner.

Finally this patch improves some comments and removes the minuses from
RSSI values as the physical sensors display the RSSI value as a positive
value.