Implement TDMA scheduler for the physical XBee sensor
This patch implements the TDMA scheduler for the physical XBee sensor.
During the implementation I encountered a silly bug in the TDMA
scheduler code that caused the timestamps to be calculated incorrectly.
This was due to integer division where float division was meant. Fixing
this issue now also makes the viewer not "jump around" and simply start
at the first node, as it should. That was bothering me for a while, but
I couldn't find out what was causing that, until now...

Furthermore this patch reduces the number of methods calls for obtaining
the setting values, which mainly makes the code more readable.