Properly destruct XBee objects
This patch fixes the hang issue that occurs when one presses Ctrl-C to
try to end the process. It turns out that the __del__ method is not
guaranteed to be called in Python, i.e., it is not a destructor, but
more a finalizer, since the `del` statement just decreases the reference
count by one.

In order to really deactivate the sensors, we implement a `deactivate()`
method and call this method upon a keyboard interrupt in the runner.

Furthermore this patch implement keyboard interrupt handling for the
XBee sensor simulator runner too.