mission_basic: Remove some try-except blocks
These are only here either to give a default value if there is a missing
settings key, or to prevent retrieving a nonexistent class from a module
(Geometry). In either case, we don't want to silently ignore the error
and start the vehicle, since that might start it in a wrong
configuration without actually knowing. Instead just raise the error and
we can fix it.