Mission: Split up into mission types
Make it possible to choose which mission to perform with a setting.
Split out certain functions from the superclass Mission into
Mission_Auto and Mission_Guided, which are "base" classes for actual
implementations for missions.

Currently supported are Mission_Square, which is an Auto mission that
flies to predetermined locations, and Mission_Browse, which stays at one
position and looks around. This makes it easier to switch between
mission types and to create new trajectory missions.