Set vehicle_simulation to ardupilot by default
The mission script does not make use of command line arguments, so we
cannot set the simulation in such a way that the vehicle is actually
from ardupilot. Meanwhile, for the standalone simulation, we can easily
use `python mission_basic.py --vehicle_simulation mock` to use the mock
vehicle. Make ardupilot the default for now.