Implement sensor address handling
This patch moves the sensor addresses to the settings file.
Additionally, it implement sensor address handling, i.e., setting the
current sensor's address and sending to all other addresses in the
network except for its own address. Furthermore the receive handler is
solidified to check which type of message is received.

Note that we do not perform a node discovery for the sensors. Node
discovery works by sending a broadcast message, which is very bad for
performance. On top of that, there is no receive acknowledgement, so we
do not know if all nodes are actually aware of the existence. Finally,
    since we have to specify the number of nodes in the network anyway,
    specifying their addresses is not unrealistic.