mission_basic: Determine simulation mode
Instead of relying on a setting, determine whether the script is being
run directly or that it is run via a module (dronekit API). This
determines whether we use mock simulation or the dronekit connections.

Add a check to stop running the code if the settings say we are not
simulating, so that this setting is still used as a safety guard.
Either way, the setting is used to determine environment simulation,
now as boolean.

For the dronekit/ArduPilot simulation, ensure that the geometry is
spherical, since we otherwise get incorrect missions that use meters
instead of coordinates. Set spherical as default since that is what we
eventually want to use too.