Arguments: Handle boolean settings
If the type of a setting is bool, then we should not rely on the default
behavior of the argument parse to handle it. This is because all
nonempty strings are True, which makes it hard to disable an option.
Instead, we register another argument for a setting x, known as --no-x,
to disable that setting, while just --x enables it.

Add tests to check that the setting is disabled when we pass the --no
version in the arguments.