mission_basic: Fix point position by angle change
Using the ArduPilot simulation and the spherical geometry, at appeared
as if points sometimes get placed at a different location than the
objects actually are. This was first considered to be a problem with the
spherical geometry, but in fact it was caused by a late request for the
vehicle's yaw.

Since in ArduPilot the yaw updates slowly (and not immediate as is the
case in the mock vehicle), the yaw may have changed between the distance
sensor request and the actual handling of its response. This is because
the angle was only requested in case we actually were going to draw the
point on the memory map (since it was otherwise useless to retrieve it).
The angle may have already changed compared to how the distance sensor
used it, which means the position is also changed compared to where the
detected object actually is.

Move the call to Distance_Sensor.get_angle to before the distance
request so that it is much more in line with the used angle. This is the
case for the simulated distance sensor, but will also have an effect for
the physical sensor since both implementations have a delay.