Implement more realistic packets for the simulator
In the packets we need the location data (latitude and longitude) of
both the sender and the receiver, as well as the RSSI value for that
link. The timestamp is only used for synchronization in the network and
can therefore be omitted when we communicate with the ground station.
This also holds for the IDs of the sensors.

Now we do not have a RSSI list anymore, but rather a data list with
custom packets (data points).

The tests have been updated accordingly. The physical sensor will be
adjusted in a follow-up commit.