Add display viewer for the environment objects
This allows opening the VRML scene using the environment and then
displaying the result using pyglet. This is very helpful in checking
whether the objects are actually positioned where they should.

Each object is drawn in a different color, and as wireframe to prevent
drawing over another. There is support for moving through the scene, but
no way to enable rotation yet.