VRMLLoader: Fix another transform bug
When a Transform node for some reason does not provide a transformation
matrix, we should just regard it as a grouping node and pass along the
transformation matrix of the parent. Because of a missing else, we
instead pass along the matrix of a previous node that has the same
parent, which causes our nodes to be displaced.

The 'castle' demo VRML file is now correctly positioned.