VRMLLoader: Add global translation
Make it possible to translate all objects in the scene, i.e., we can
have the vehicle start at a different location than the scene's center.
Make this a setting of a three-element list so that it can be passed
from the command line as an argument with 1 through 3 values.