Keep track of frame IDs to match packets
The main flaw with the previous version was that we did not attach a
unique identifier to the packet for requesting the RSSI. If we then
received an RSSI packet we did not know to which original packet it
belongs, so we just assumed the last one in the list. This, however,
cannot be assumed due to the asynchronicity of the system.

This patch:
- Links a received packet to a unique frame ID. The frame ID is only one
  byte, so we cannot be too unique, however this is good enough for our
  purpose.
- Adds the frame ID to the RSSI request.
- Matches a received RSSI packet with a received original packet using
  the frame IDs.
- Sends only those packets to the ground sensor that have a
  corresponding RSSI value. Otherwise, the packet is invalid and
  therefore dropped.

Additionally, this patch:
- Adds a note about documentation.
- Removes a note about the intermittent list failure as that is now
  fixed.
- Updates the settings to set more realistic values for our setup.
- Updates the tests to use custom settings as we do not want the viewer
  tests to rely on the production settings file.