Mission_Search: Initial attempt
Use Mission_Browse's stepwise rotation checks to deduce what the
"safest" direction (that at least has something in its path) is. Once
we're done scanning, move in that direction and check again once we're
getting close.

This works mostly, even though it is quite naive to move in these
directions that way.

Fix the MockVehicle to handle Mission.set_speed in non-AUTO mode as
well, since that is apparently possible and also helpful in this case,
rather than calculating the global velocity ourselves.