VRMLLoader: Replace OpenGLContext with vrml parser
Use the VRML parser ourselves rather than rely on OpenGLContext for it.
This removes a dependency on a packages that is hard to set up.
OpenGLContext is helpful if we would want to read other types of files,
but we can just use Blender to convert such files to VRML.