Viewer: Fix OpenGL coordinates again
Apparently the whole scene was being drawn mirrored because the latitude
or y axis should actually be drawn inward rather than outward (north).
This causes the viewer to draw an object that appeared to be far away
for both the distance calculation and the viewer itself, even though it
was actually the object that is supposed in front of us. By flipping the
latitude value before drawing, we now should be conforming to default GL
geometry as well as correctly detect the objects.

Note that this does make the viewer look southward instead of northward
initially. According to the VRML specification, this might be standard:
http://www.web3d.org/documents/specifications/14772/V2.0/part1/concepts.html#4.4.5
although "on the Z-axis looking down the -Z-axis toward the origin" is
a somewhat ambiguous phrasing to me.