Distance_Sensor_Simulator: Plane detection in spherical geometry
Make use of Location objects and differences given by the Geometry
object, such that the code becomes compatible with Geometry_Spherical.
This causes objects to actually be detected when geographic coordinates
are used. Note that this is not yet visible in the viewer, which is not
compatible with spherical geometry right now.