Monitor: Add setting to skip memory map plot
Allow giving the argument --no-plot to stop any plotting from happening.
This will probably have to become the default when doing physical runs.

Also skip plotting simulated polygons when the environment is not
simulated. This also has to be done for the current sensor edge, which
the physical distance sensor will not implement, but that can follow
later.