Monitor: Display angle in memory map
When we are plotting the memory map, display not only the vehicle
location but also its angle with a red arrow pointing in the 2D
direction, like the ArduPilot simulator does. In combination with the
display viewer, this provides a lot of information.

Add a helpful method for getting the vehicle angle rather than the yaw,
leaving conversion to be done in the Environment.