Geometry: Add angle difference check method
Rather than using diff_angle, making at absolute and comparing it
against a margin ourselves in various classes, use a new method
Geometry.check_angle to do all this, consistently. Use less-or-equal
rather than less-than to be inside the allowed difference margin since
that makes more sense when a difference on 0 radians is given.

Add tests for the new method.

Clean up the usages and some prints in MockVehicle as well.