Viewer: Fix redraw on resize
For some reason the window is redrawn when there are two event handlers
(the latter being the window event logger), but it is not when we only
have our own handler. This causes the viewer to not display the scene
immediately at start and also not work correctly on resize.

Invalidate the window so that it is redrawn afterward.