Vewer: Simplify rotation
Since we only use pitch and yaw rotation and have three dimensions to
work with, we can actually easily map these to two of the dimensions and
leave the other one as is. Fix the rotation matrix before drawing so
that the dimensions map correctly, and remove the whole rotation
reorientation.