Viewer: Extend rotation with camera vectors
Tracking the rotation angles themselves is not enough, because this
causes one of the rotations to not work correctly when the other one
rotates (i.e., the pitch becomes a roll rotation if we rotate yaw, or
the yaw becomes roll when pitch is rotated).

To avoid this, make use of vectors that determine the position of the
camera, the point it is looking at, and "up" and "right" vectors. This
allows free rotation in all axes without so-called gimbal lock. All
rotation and movement updates are now matrix and vector multiplications.
This also makes use of gluLookAt camera positioning instead of
transforming the world, which is conceptually better. There is more room
for errors with all these vectors, also due to the conversion between
OpenGL and lat/lon/alt coordinate systems, but fixing it is worth it and
everything should work as expected, but strafing now follows camera
orientation.

Add a Vector class which is a numpy array with nice accessors to make
code more readable.