Viewer: Fix z axis in/out movement
The W key is for inward and the S key is for outward in the WASD key
scheme. This was accomplished by negating both the Z axis strafe and the
final Z axis move after rotation. This caused the W and S keys to do
something different when we rotate, e.g. when viewing upward: the W then
goes down and S goes up, which is counterintuitive. Furthermore, this
removes a few unexplainable negations so that things make more sense.