Geometry: Fix up pitch calculations
With pitch, yaw and roll rotation, the order in which the rotations are
applied is roll, pitch, then yaw, where we do not support roll so far.
However, when we want to use rotation angles to calculate the point at
a certain distance, we have to project the frame by applying the angles
in the reverse order. Otherwise we cannot make use of the pitch rotation
before we calculate everything for the yaw rotation. This restores the

Additionally:
- Use altitude difference in get_distance_meters, this gave wrong
  distances when looking up before. Make sure that spherical geometry,
  where altitude is the only non-geocoordinate value, applies the
  distances correctly with the coordinate to meters conversion.
- Remove use of get_angle in Distance_Sensor_Simulator, since it only
  works for 2D (ignoring altitude) and it is better to check whether the
  factor is nonnegative in order to find out if the detected
  intersection point is in front of us instead of back of us.
- Tests: Separate deltas for coordinate differences from distance
  differences, otherwise the spherical geometry gets too much leeway in
  certain tests.