Viewer: Fix pitch and yaw calculation
The rotation of the camera is such that if we are facing backward (yaw),
then rotating the camera upward (pitch) does indeed move the camera up,
which is nice. However, the pitch itself should decrease, since the
pitch is normally applied before the yaw and thus we should be looking
down when we would face forward. Calculate the pitch and yaw as
described in http://www.jldoty.com/code/DirectX/YPRfromUF/YPRfromUF.html
making use of the camera vectors to find the actual values. The
interactive viewer can now detect objects in all directions without any
problem.