Mark all internal packet items as private
Prepend an underscore for all internal packet items to signal that they
should not be added, removed or changed by code not belonging to the
framework's core.

Additionally update a unit test that for some reason succeeded even
though it had not been updated to use the new packet structure.