Swap custom packet transmission and ground station transmission
This makes the physical XBee code stable. I'm not exactly sure as to why
as it should be equivalent, but I assume that the ground sensor loop
takes too much time (because there might also be many packets that are
not ready to be sent yet), causing the custom packets to be dropped.