Servo: Initial servo with distance sensor support
Make it possible to define servos that can be controlled via the flight
controller. The pin number and minimum/maximum angles and PWMs can be
configured. We can retrieve the current servo values via regular
messages from the controller, and change servos depending on which angle
we wish to see from within the mission.

The distance sensors now have IDs to make it easier to track which one
they are, and correspond to servos. They can request the current servo
angle from the environment and use this for the simulated calculations.