MockVehicle: Track home location from translation
When we use a simulated scene, we can translate the entire scene to
simulate starting from a different location. Let the mock vehicle know
the concept of a home location, somewhat like actual vehicle does
(although we cannot control that). We use this translation to determine
the home location so that they are different for drones starting at
different locations.

Fix the use of the translation in VRMLLoader, where the longitude should
be altered in another direction to make sense, apparently.