Monitor: Share memory map updates with each other
This allows the vehicle instances to send a packet about a location that
has a detected object, so that the other vehicles know about these
obstacles. This makes use of the home location to convert the relative
location so that it can be used on other vehicles that have a different
home location.

This does not work perfectly yet with the locations, but the custom
packets are received and added to the memory map just fine.