Geometry: Equalize relativeness with home location
If one of the locations given to geometry methods is not relative while
the other one is, then calculations with the raw lat/lon/alt information
is incorrect. Ensure that in this case both are absolute coordinates.
Allow giving Geometry a home location, updated by the environment, and
use this to fix the location relativeness.