Monitor: Disable XBee in development by default
We are developing other features as well, and having another thread
interferes with error handling somewhat. Thus only start a thread for
the XBee sensor when we use the sensors to prevent an idle thread, and
set the type to nothing in the settings defaults since it is nicer to
pass it in the arguments than it is to clear it.