Mission: Add waypoint caching in Mission_Auto
Do not call get_points every time in Mission_Square, instead move the
implementation of add_commands to Mission_Auto and add caching around of
get_waypoints. Make get_points a method that should be implemented.

This prevents creating a bunch of locations every time we draw the
memory map.