Distance_Sensor_Simulator: Fix detection
For the ray intersection, we need a second point on the line so that we
can get an equation of the line (u). We used epsilon for this since it
seemed as if closer points worked better, but that was in an earlier
buggy implementation. Now it seems that certain objects do not get
detected because of the dot product being too small, simply due to the
difference not contributing enough. Use a greater distance for the
second point, so that all parts of the trees in trees_river.wrl are
detected.